+++ examples/sequence-diagrams-with-sqlite-database/api_test.go
@@ -90,7 +90,7 @@ func postUserMock(username string) *spectest.Mock {
 		End()
 }
 
-func apiTest(name string) *spectest.APITest {
+func apiTest(name string) *spectest.SpecTest {
 	dsn := os.Getenv("SQLITE_DSN")
 
 	// Connect using the previously registered driver